lambda/requirements.txt
# Core dependencies
requests==2.32.3
python-dotenv==1.0.1
beautifulsoup4==4.12.3

# Database
psycopg2-binary==2.9.11

# LangChain ecosystem
langchain==0.3.14
langchain-core==0.3.27
langchain-openai==0.2.14
langchain-anthropic==0.3.8
langgraph==0.2.58

# AWS
boto3==1.35.96
botocore==1.35.96

# Additional utilities
numpy==1.26.4
pydantic==2.10.6
tenacity==9.0.0
